3033 - rename naming conventions.
