+++ wls-gui-wahllokalsystem/src/components/common/widgets/BaseOfflineDataSyncWidget.vue
